# optimize transitions