master:
```
julia> @benchmark parse_fasta("""
       >NP_003172.1 brachyury protein isoform 1 [Homo sapiens]
       MSSPGTESAGKSLQYRVDHLLSAVENELQAGSEKGDPTERELRVGLEESELWLRFKELTNEMIVTKNGRR
       MFPVLKVNVSGLDPNAMYSFLLDFVAADNHRWKYVNGEWVPGGKPEPQAPSCVYIHPDSPNFGAHWMKAP
       VSFSKVKLTNKLNGGGQIMLNSLHKYEPRIHIVRVGGPQRMITSHCFPETQFIAVTAYQNEEITALKIKY
       NPFAKAFLDAKERSDHKEMMEEPGDSQQPGYSQWGWLLPGTSTLCPPANPHPQFGGALSLPSTHSCDRYP
       TLRSHRSSPYPSPYAHRNNSPTYSDNSPACLSMLQSHDNWSSLGMPAHPSMLPVSHNASPPTSSSQYPSL
       WSVSNGAVTPGSQAAAVSNGLGAQFFRGSPAHYTPLTHPVSAPSSSGSPLYEGAAAATDIVDSQYDAAAQ
       GRLIASWTPVSPPSM
       """)
BenchmarkTools.Trial:
  memory estimate:  1.78 KiB
  allocs estimate:  14
  --------------
  minimum time:     1.643 μs (0.00% GC)
  median time:      1.768 μs (0.00% GC)
  mean time:        2.058 μs (8.76% GC)
  maximum time:     250.817 μs (98.33% GC)
  --------------
  samples:          10000
  evals/sample:     10
  time tolerance:   5.00%
  memory tolerance: 1.00%

```

This branch:
```
julia> @benchmark parse_fasta("""
       >NP_003172.1 brachyury protein isoform 1 [Homo sapiens]
       MSSPGTESAGKSLQYRVDHLLSAVENELQAGSEKGDPTERELRVGLEESELWLRFKELTNEMIVTKNGRR
       MFPVLKVNVSGLDPNAMYSFLLDFVAADNHRWKYVNGEWVPGGKPEPQAPSCVYIHPDSPNFGAHWMKAP
       VSFSKVKLTNKLNGGGQIMLNSLHKYEPRIHIVRVGGPQRMITSHCFPETQFIAVTAYQNEEITALKIKY
       NPFAKAFLDAKERSDHKEMMEEPGDSQQPGYSQWGWLLPGTSTLCPPANPHPQFGGALSLPSTHSCDRYP
       TLRSHRSSPYPSPYAHRNNSPTYSDNSPACLSMLQSHDNWSSLGMPAHPSMLPVSHNASPPTSSSQYPSL
       WSVSNGAVTPGSQAAAVSNGLGAQFFRGSPAHYTPLTHPVSAPSSSGSPLYEGAAAATDIVDSQYDAAAQ
       GRLIASWTPVSPPSM
       """)
BenchmarkTools.Trial:
  memory estimate:  1.78 KiB
  allocs estimate:  14
  --------------
  minimum time:     1.232 μs (0.00% GC)
  median time:      1.310 μs (0.00% GC)
  mean time:        1.559 μs (11.11% GC)
  maximum time:     240.000 μs (98.20% GC)
  --------------
  samples:          10000
  evals/sample:     10
  time tolerance:   5.00%
  memory tolerance: 1.00%

```